requirements.txt
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-community==0.3.26
chromadb==0.4.23
openai==1.91.0
tiktoken==0.7.0
pytesseract==0.3.10
pdfplumber==0.10.3
python-dotenv==1.0.1
streamlit==1.35.0
Pillow==10.3.0
